refactor: 마이페이지 랭킹 응답 DTO를 로컬 정의로 변경
