+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/member/controller/MyPageController.java
@@ -3,11 +3,11 @@ package me.bombom.api.v1.member.controller;
 import lombok.RequiredArgsConstructor;
 import me.bombom.api.v1.common.resolver.LoginMember;
 import me.bombom.api.v1.member.domain.Member;
+import me.bombom.api.v1.member.dto.response.RankSummaryResponse;
 import me.bombom.api.v1.member.service.MemberService;
 import me.bombom.api.v1.member.service.MyPageService;
 import me.bombom.openapi.mypage.api.MyPageApi;
 import me.bombom.openapi.mypage.model.MemberJoinDaysResponse;
-import me.bombom.openapi.mypage.model.RankSummaryResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -27,7 +27,6 @@ public class MyPageController implements MyPageApi {
         return memberService.getJoinDays(member);
     }
 
-    @Override
     @GetMapping("/rank")
     public RankSummaryResponse getRankSummary(
             @LoginMember Member member,
