+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/member/dto/response/RankSummaryResponse.java
@@ -1,0 +1,12 @@
+package me.bombom.api.v1.member.dto.response;
+
+import java.util.List;
+
+public record RankSummaryResponse(
+        List<RankCardResponse> cards
+) {
+
+    public static RankSummaryResponse from(List<RankCardResponse> cards) {
+        return new RankSummaryResponse(cards);
+    }
+}
