+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/member/service/MyPageService.java
@@ -10,6 +10,9 @@ import me.bombom.api.v1.common.exception.CIllegalArgumentException;
 import me.bombom.api.v1.common.exception.ErrorContextKeys;
 import me.bombom.api.v1.common.exception.ErrorDetail;
 import me.bombom.api.v1.member.domain.Member;
+import me.bombom.api.v1.member.dto.response.RankCardResponse;
+import me.bombom.api.v1.member.dto.response.RankHistoryResponse;
+import me.bombom.api.v1.member.dto.response.RankSummaryResponse;
 import me.bombom.api.v1.member.enums.MyPageRankType;
 import me.bombom.api.v1.reading.domain.ContinueReadingRankHistory;
 import me.bombom.api.v1.reading.domain.ContinueReadingRealtime;
@@ -17,9 +20,6 @@ import me.bombom.api.v1.reading.domain.MonthlyReadingRankHistory;
 import me.bombom.api.v1.reading.repository.ContinueReadingRankHistoryRepository;
 import me.bombom.api.v1.reading.repository.ContinueReadingRealtimeRepository;
 import me.bombom.api.v1.reading.repository.MonthlyReadingRankHistoryRepository;
-import me.bombom.openapi.mypage.model.RankCardResponse;
-import me.bombom.openapi.mypage.model.RankHistoryResponse;
-import me.bombom.openapi.mypage.model.RankSummaryResponse;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
